portfolio/requirements.txt
streamlit
streamlit-option-menu
python-dotenv
pandas
numpy
requests
certifi
matplotlib
plotly
beautifulsoup4
konlpy
selenium
google-auth
google-api-python-client
google-genai  # Note: Confirm the official package name for Google Gemini AI SDK
youtube-transcript-api
stqdm
pygsheets
